refactor(common): split out URL-related code into UrlService
Move code from Location to UrlService. Move tests from location_spec to url_service_spec for behavior now in UrlService.

This is part of an effort to allow the Angular Router to use the new [Navigation API](https://developer.mozilla.org/en-US/docs/Web/API/Navigation_API) instead of the History API(https://developer.mozilla.org/en-US/docs/Web/API/History_API).

The relevant part of that effort is to introduce a new public API that can use either Location or a new NavigationService.
